% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/board.R
\name{board}
\alias{board}
\title{R6 class storing health statistics for a health board.}
\description{
R6 class storing health statistics for a health board.

R6 class storing health statistics for a health board.
}
\section{Super class}{
\code{\link[HealthDataScotland:health_unit]{HealthDataScotland::health_unit}} -> \code{board}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-board-available_plots}{\code{board$available_plots()}}
\item \href{#method-board-plot}{\code{board$plot()}}
\item \href{#method-board-address}{\code{board$address()}}
\item \href{#method-board-clone}{\code{board$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="HealthDataScotland" data-topic="health_unit" data-id="id"><a href='../../HealthDataScotland/html/health_unit.html#method-health_unit-id'><code>HealthDataScotland::health_unit$id()</code></a></li>
<li><span class="pkg-link" data-pkg="HealthDataScotland" data-topic="health_unit" data-id="initialize"><a href='../../HealthDataScotland/html/health_unit.html#method-health_unit-initialize'><code>HealthDataScotland::health_unit$initialize()</code></a></li>
<li><span class="pkg-link" data-pkg="HealthDataScotland" data-topic="health_unit" data-id="title"><a href='../../HealthDataScotland/html/health_unit.html#method-health_unit-title'><code>HealthDataScotland::health_unit$title()</code></a></li>
<li><span class="pkg-link" data-pkg="HealthDataScotland" data-topic="health_unit" data-id="validate"><a href='../../HealthDataScotland/html/health_unit.html#method-health_unit-validate'><code>HealthDataScotland::health_unit$validate()</code></a></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-board-available_plots"></a>}}
\if{latex}{\out{\hypertarget{method-board-available_plots}{}}}
\subsection{Method \code{available_plots()}}{
Get character vector of available plots for board unit. Options
are either "specialty_bar" plot.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{board$available_plots()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-board-plot"></a>}}
\if{latex}{\out{\hypertarget{method-board-plot}{}}}
\subsection{Method \code{plot()}}{
Plot board unit.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{board$plot(type, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{type}}{(character(1))\cr
Character specifying plot type. See \code{available_plots} for options.}

\item{\code{...}}{Passed to plot functions.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-board-address"></a>}}
\if{latex}{\out{\hypertarget{method-board-address}{}}}
\subsection{Method \code{address()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{board$address()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-board-clone"></a>}}
\if{latex}{\out{\hypertarget{method-board-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{board$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
