% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_data.R
\docType{data}
\name{example_board_data}
\alias{example_board_data}
\title{Example health board dataset.}
\description{
Data.frame object containing specialty bed data for Scottish health boards
for 2022/24.
}
\seealso{
Other health data scotland examples: 
\code{\link{example_board_metadata}},
\code{\link{example_gp_data}},
\code{\link{example_gp_metadata}},
\code{\link{example_hospital_data}},
\code{\link{example_hospital_metadata}}
}
\concept{health data scotland examples}
\keyword{data}
