% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_data.R
\docType{data}
\name{example_board_metadata}
\alias{example_board_metadata}
\title{Example health board metadata.}
\description{
Data.frame object containing metadata for scottish health boards.
}
\seealso{
Other health data scotland examples: 
\code{\link{example_board_data}},
\code{\link{example_gp_data}},
\code{\link{example_gp_metadata}},
\code{\link{example_hospital_data}},
\code{\link{example_hospital_metadata}}
}
\concept{health data scotland examples}
\keyword{data}
