% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gp_grp.R
\name{example_gp_grp_unit}
\alias{example_gp_grp_unit}
\title{Get example gp health unit grp object.}
\usage{
example_gp_grp_unit(ids = c("10002", "10017"))
}
\arguments{
\item{ids}{Character IDs of GP practice to get. Default is c("10002", "10017").}
}
\description{
Get example gp health unit grp object.
}
