% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_data.R
\docType{data}
\name{example_gp_metadata}
\alias{example_gp_metadata}
\title{Example GP metadata.}
\description{
Data.frame object containing metadata for scottish GP practices
from April 2024.
}
\seealso{
Other health data scotland examples: 
\code{\link{example_board_sf}},
\code{\link{example_gp_data}},
\code{\link{example_gp_sf}},
\code{\link{example_hospital_data}},
\code{\link{example_hospital_metadata}},
\code{\link{example_hospital_sf}}
}
\concept{health data scotland examples}
\keyword{data}
