% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gp.R
\name{example_gp_unit}
\alias{example_gp_unit}
\title{Get example gp health unit object.}
\usage{
example_gp_unit(id = "10002")
}
\arguments{
\item{id}{Character ID of GP practice to get. Default is "10002".}
}
\description{
Get example gp health unit object.
}
