% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hospital_grp.R
\name{example_hospital_grp_unit}
\alias{example_hospital_grp_unit}
\title{Get example gp health unit grp object.}
\usage{
example_hospital_grp_unit(ids = c("A101H", "A201H"))
}
\arguments{
\item{ids}{Character ID of Hospital to get. Default is c("A101H", "A201H").}
}
\description{
Get example gp health unit grp object.
}
