% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_data.R
\docType{data}
\name{example_hospital_sf}
\alias{example_hospital_sf}
\title{Example hospital spatial data}
\description{
sf data.frame object containing spatial data for hospitals in Scotland as
of 2021. See \code{inst/extdata/READM.md} for details.
}
\seealso{
Other health data scotland examples: 
\code{\link{example_board_sf}},
\code{\link{example_gp_data}},
\code{\link{example_gp_metadata}},
\code{\link{example_gp_sf}},
\code{\link{example_hospital_data}},
\code{\link{example_hospital_metadata}}
}
\concept{health data scotland examples}
\keyword{data}
