% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hospital.R
\name{example_hospital_unit}
\alias{example_hospital_unit}
\title{Get example hospital health unit object.}
\usage{
example_hospital_unit(id = "A101H")
}
\arguments{
\item{id}{Character ID of Hospital to get. Default is "A101H".}
}
\description{
Get example hospital health unit object.
}
