% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_download.R
\name{get_sf}
\alias{get_sf}
\title{Get SF spatial data.frame for selected data type.}
\usage{
get_sf(type = c("gp", "hospital", "board"))
}
\arguments{
\item{type}{Character type of spatial data. Options are
"gp", "hospital", or "board".}
}
\description{
Get SF spatial data.frame for selected data type.
}
