% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/health_data_scotland.R
\name{health_data_scotland}
\alias{health_data_scotland}
\title{Launch shiny application}
\usage{
health_data_scotland(...)
}
\arguments{
\item{...}{Passed to shiny::shinyApp.}
}
\description{
Launch shiny application
}
