% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/health_unit.R
\name{health_unit}
\alias{health_unit}
\title{R6 class representing a health unit e.g. a GP practice or hospital.}
\description{
R6 class representing a health unit e.g. a GP practice or hospital.

R6 class representing a health unit e.g. a GP practice or hospital.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{.metadata}}{A data.frame storing meatadata for health unit}

\item{\code{.data}}{A data.frame storing health statistics for health unit.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-health_unit-new}{\code{health_unit$new()}}
\item \href{#method-health_unit-validate}{\code{health_unit$validate()}}
\item \href{#method-health_unit-metadata}{\code{health_unit$metadata()}}
\item \href{#method-health_unit-data}{\code{health_unit$data()}}
\item \href{#method-health_unit-combine_data}{\code{health_unit$combine_data()}}
\item \href{#method-health_unit-ID}{\code{health_unit$ID()}}
\item \href{#method-health_unit-title}{\code{health_unit$title()}}
\item \href{#method-health_unit-address}{\code{health_unit$address()}}
\item \href{#method-health_unit-health_board}{\code{health_unit$health_board()}}
\item \href{#method-health_unit-popup_modal}{\code{health_unit$popup_modal()}}
\item \href{#method-health_unit-clone}{\code{health_unit$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-health_unit-new"></a>}}
\if{latex}{\out{\hypertarget{method-health_unit-new}{}}}
\subsection{Method \code{new()}}{
Create instance of health unit.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{health_unit$new(.metadata, .data)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{.metadata}}{(\code{data.frame})\cr
A data.frame storing metadata for health unit.}

\item{\code{.data}}{(\code{data.frame})\cr
A data.frame storing health statistics for health unit.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-health_unit-validate"></a>}}
\if{latex}{\out{\hypertarget{method-health_unit-validate}{}}}
\subsection{Method \code{validate()}}{
Validate structure of health unit.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{health_unit$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-health_unit-metadata"></a>}}
\if{latex}{\out{\hypertarget{method-health_unit-metadata}{}}}
\subsection{Method \code{metadata()}}{
Get metadata of health unit.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{health_unit$metadata()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-health_unit-data"></a>}}
\if{latex}{\out{\hypertarget{method-health_unit-data}{}}}
\subsection{Method \code{data()}}{
Get data of health unit.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{health_unit$data()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-health_unit-combine_data"></a>}}
\if{latex}{\out{\hypertarget{method-health_unit-combine_data}{}}}
\subsection{Method \code{combine_data()}}{
Get combined metadata and data in single data.frame.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{health_unit$combine_data()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-health_unit-ID"></a>}}
\if{latex}{\out{\hypertarget{method-health_unit-ID}{}}}
\subsection{Method \code{ID()}}{
Get ID of health unit.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{health_unit$ID()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-health_unit-title"></a>}}
\if{latex}{\out{\hypertarget{method-health_unit-title}{}}}
\subsection{Method \code{title()}}{
Get title of health unit.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{health_unit$title()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-health_unit-address"></a>}}
\if{latex}{\out{\hypertarget{method-health_unit-address}{}}}
\subsection{Method \code{address()}}{
Get address of health unit.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{health_unit$address()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-health_unit-health_board"></a>}}
\if{latex}{\out{\hypertarget{method-health_unit-health_board}{}}}
\subsection{Method \code{health_board()}}{
Get health board of health unit.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{health_unit$health_board()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-health_unit-popup_modal"></a>}}
\if{latex}{\out{\hypertarget{method-health_unit-popup_modal}{}}}
\subsection{Method \code{popup_modal()}}{
Initialise model popup of health unit.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{health_unit$popup_modal(ns)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{ns}}{(character(1))\cr
Namespace to create health unit UI.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-health_unit-clone"></a>}}
\if{latex}{\out{\hypertarget{method-health_unit-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{health_unit$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
