% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hospital.R
\name{hospital}
\alias{hospital}
\title{R6 class storing health statistics for a single hospital.}
\description{
R6 class storing health statistics for a single hospital.

R6 class storing health statistics for a single hospital.
}
\details{
This R6 class is designed to store bed occupancy statistics for a single
hospital location. This class can be used to plot summary statistics and
create shiny UI/server objects.
}
\examples{
library(dplyr)
meta <- HealthDataScotland::example_hospital_metadata |>
  rename("ID" = "HospitalCode", "HBName" = "HealthBoard") |>
  filter(.data[["ID"]] == "A101H")
data <- HealthDataScotland::example_hospital_data |>
  rename("ID" = "Location") |>
  filter(.data[["ID"]] == "A101H")
x <- hospital[["new"]](meta, data)
x[["ID"]]()
x[["title"]]()
x[["address"]]()
x[["health_board"]]()
x[["metadata"]]()
x[["data"]]()
x[["plot"]](type = "specialty_line")
x[["plot_data"]](type = "specialty_line")
x[["plot_info"]](type = "specialty_line")
x[["summary"]](type = "specialty_summary")
x[["summary_info"]](type = "specialty_summary")
\dontrun{
x[["ui"]]()
x[["server"]]()
}

## ------------------------------------------------
## Method `hospital$plot`
## ------------------------------------------------

x <- example_hospital_unit()
x[["plot"]](type = "specialty_line", data_type = "annual")

## ------------------------------------------------
## Method `hospital$plot_data`
## ------------------------------------------------

x <- example_hospital_unit()
x[["plot_data"]](type = "specialty_line", data_type = "annual")

## ------------------------------------------------
## Method `hospital$plot_info`
## ------------------------------------------------

x <- example_hospital_unit()
x[["plot_info"]](type = "specialty_line", data_type = "annual")

## ------------------------------------------------
## Method `hospital$summary`
## ------------------------------------------------

x <- example_hospital_unit()
x[["summary"]](type = "specialty_summary")

## ------------------------------------------------
## Method `hospital$summary_info`
## ------------------------------------------------

x <- example_hospital_unit()
x[["summary_info"]](type = "specialty_summary")
}
\section{Super class}{
\code{\link[HealthDataScotland:health_unit]{HealthDataScotland::health_unit}} -> \code{hospital}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-hospital-plot_types}{\code{hospital$plot_types()}}
\item \href{#method-hospital-plot}{\code{hospital$plot()}}
\item \href{#method-hospital-plot_data}{\code{hospital$plot_data()}}
\item \href{#method-hospital-plot_info}{\code{hospital$plot_info()}}
\item \href{#method-hospital-summary_types}{\code{hospital$summary_types()}}
\item \href{#method-hospital-summary}{\code{hospital$summary()}}
\item \href{#method-hospital-summary_info}{\code{hospital$summary_info()}}
\item \href{#method-hospital-ui}{\code{hospital$ui()}}
\item \href{#method-hospital-server}{\code{hospital$server()}}
\item \href{#method-hospital-clone}{\code{hospital$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="HealthDataScotland" data-topic="health_unit" data-id="ID"><a href='../../HealthDataScotland/html/health_unit.html#method-health_unit-ID'><code>HealthDataScotland::health_unit$ID()</code></a></span></li>
<li><span class="pkg-link" data-pkg="HealthDataScotland" data-topic="health_unit" data-id="address"><a href='../../HealthDataScotland/html/health_unit.html#method-health_unit-address'><code>HealthDataScotland::health_unit$address()</code></a></span></li>
<li><span class="pkg-link" data-pkg="HealthDataScotland" data-topic="health_unit" data-id="combine_data"><a href='../../HealthDataScotland/html/health_unit.html#method-health_unit-combine_data'><code>HealthDataScotland::health_unit$combine_data()</code></a></span></li>
<li><span class="pkg-link" data-pkg="HealthDataScotland" data-topic="health_unit" data-id="data"><a href='../../HealthDataScotland/html/health_unit.html#method-health_unit-data'><code>HealthDataScotland::health_unit$data()</code></a></span></li>
<li><span class="pkg-link" data-pkg="HealthDataScotland" data-topic="health_unit" data-id="datatable"><a href='../../HealthDataScotland/html/health_unit.html#method-health_unit-datatable'><code>HealthDataScotland::health_unit$datatable()</code></a></span></li>
<li><span class="pkg-link" data-pkg="HealthDataScotland" data-topic="health_unit" data-id="health_board"><a href='../../HealthDataScotland/html/health_unit.html#method-health_unit-health_board'><code>HealthDataScotland::health_unit$health_board()</code></a></span></li>
<li><span class="pkg-link" data-pkg="HealthDataScotland" data-topic="health_unit" data-id="initialize"><a href='../../HealthDataScotland/html/health_unit.html#method-health_unit-initialize'><code>HealthDataScotland::health_unit$initialize()</code></a></span></li>
<li><span class="pkg-link" data-pkg="HealthDataScotland" data-topic="health_unit" data-id="metadata"><a href='../../HealthDataScotland/html/health_unit.html#method-health_unit-metadata'><code>HealthDataScotland::health_unit$metadata()</code></a></span></li>
<li><span class="pkg-link" data-pkg="HealthDataScotland" data-topic="health_unit" data-id="popup_modal"><a href='../../HealthDataScotland/html/health_unit.html#method-health_unit-popup_modal'><code>HealthDataScotland::health_unit$popup_modal()</code></a></span></li>
<li><span class="pkg-link" data-pkg="HealthDataScotland" data-topic="health_unit" data-id="title"><a href='../../HealthDataScotland/html/health_unit.html#method-health_unit-title'><code>HealthDataScotland::health_unit$title()</code></a></span></li>
<li><span class="pkg-link" data-pkg="HealthDataScotland" data-topic="health_unit" data-id="validate"><a href='../../HealthDataScotland/html/health_unit.html#method-health_unit-validate'><code>HealthDataScotland::health_unit$validate()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-hospital-plot_types"></a>}}
\if{latex}{\out{\hypertarget{method-hospital-plot_types}{}}}
\subsection{Method \code{plot_types()}}{
Get character vector of plot types for hospital grp
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{hospital$plot_types()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-hospital-plot"></a>}}
\if{latex}{\out{\hypertarget{method-hospital-plot}{}}}
\subsection{Method \code{plot()}}{
Plot hospital unit.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{hospital$plot(type, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{type}}{(character(1))\cr
Character specifying plot type. See \code{plot_types} for options.}

\item{\code{...}}{Passed to plot functions.}
}
\if{html}{\out{</div>}}
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{x <- example_hospital_unit()
x[["plot"]](type = "specialty_line", data_type = "annual")
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-hospital-plot_data"></a>}}
\if{latex}{\out{\hypertarget{method-hospital-plot_data}{}}}
\subsection{Method \code{plot_data()}}{
Generate plot data for hospital unit.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{hospital$plot_data(type, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{type}}{(character(1))\cr
Character specifying plot type. See \code{plot_types} for options.}

\item{\code{...}}{Passed to plot data functions.}
}
\if{html}{\out{</div>}}
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{x <- example_hospital_unit()
x[["plot_data"]](type = "specialty_line", data_type = "annual")
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-hospital-plot_info"></a>}}
\if{latex}{\out{\hypertarget{method-hospital-plot_info}{}}}
\subsection{Method \code{plot_info()}}{
Get plot info for hospital unit.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{hospital$plot_info(type, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{type}}{(character(1))\cr
Character specifying plot type. See \code{plot_types} for options.}

\item{\code{...}}{Passed to plot info functions.}
}
\if{html}{\out{</div>}}
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{x <- example_hospital_unit()
x[["plot_info"]](type = "specialty_line", data_type = "annual")
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-hospital-summary_types"></a>}}
\if{latex}{\out{\hypertarget{method-hospital-summary_types}{}}}
\subsection{Method \code{summary_types()}}{
Get character vector of summary types for hospital unit.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{hospital$summary_types()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-hospital-summary"></a>}}
\if{latex}{\out{\hypertarget{method-hospital-summary}{}}}
\subsection{Method \code{summary()}}{
Summarise hospital data.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{hospital$summary(type, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{type}}{(character(1))\cr
Character specifying summary type. See \code{summary_types} for options.}

\item{\code{...}}{Passed to method.}
}
\if{html}{\out{</div>}}
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{x <- example_hospital_unit()
x[["summary"]](type = "specialty_summary")
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-hospital-summary_info"></a>}}
\if{latex}{\out{\hypertarget{method-hospital-summary_info}{}}}
\subsection{Method \code{summary_info()}}{
Get summary info for hospital unit.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{hospital$summary_info(type, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{type}}{(character(1))\cr
Character specifying summary type. See \code{summary_types} for options.}

\item{\code{...}}{Passed to summary info functions.}
}
\if{html}{\out{</div>}}
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{x <- example_hospital_unit()
x[["summary_info"]](type = "specialty_summary")
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-hospital-ui"></a>}}
\if{latex}{\out{\hypertarget{method-hospital-ui}{}}}
\subsection{Method \code{ui()}}{
Create UI for hospital object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{hospital$ui(ns)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{ns}}{Namespace of shiny application page.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-hospital-server"></a>}}
\if{latex}{\out{\hypertarget{method-hospital-server}{}}}
\subsection{Method \code{server()}}{
Create server for hospital object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{hospital$server()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-hospital-clone"></a>}}
\if{latex}{\out{\hypertarget{method-hospital-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{hospital$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
