% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hospital_grp.R
\name{hospital_grp}
\alias{hospital_grp}
\title{R6 class storing health statistics for a list of hospital health units.}
\description{
R6 class storing health statistics for a list of hospital health units.

R6 class storing health statistics for a list of hospital health units.
}
\details{
This R6 class is designed to store bed occupancy statistics for multiple
hospitals. This class can be used to plot summary statistics and
create shiny UI/server objects.
}
\examples{
gps <- lapply(c("A101H", "A201H"), example_hospital_unit)
x <- hospital_grp[["new"]](gps, .id = "gp")
x[["ID"]]()
x[["IDs"]]()
x[["titles"]]()
x[["metadata"]]()
x[["data"]]()
x[["subset"]](id = "A201H")
x[["plot"]](type = "hospital_bar")
x[["plot_data"]](type = "hospital_bar")
x[["plot_info"]](type = "hospital_bar")
\dontrun{
x[["ui"]]()
x[["server"]]()
}

## ------------------------------------------------
## Method `hospital_grp$plot`
## ------------------------------------------------

x <- example_hospital_grp_unit()
x[["plot"]](type = "hospital_bar")

## ------------------------------------------------
## Method `hospital_grp$plot_data`
## ------------------------------------------------

x <- example_hospital_grp_unit()
x[["plot_data"]](type = "hospital_bar")

## ------------------------------------------------
## Method `hospital_grp$plot_info`
## ------------------------------------------------

x <- example_hospital_grp_unit()
x[["plot_info"]](type = "hospital_bar")

## ------------------------------------------------
## Method `hospital_grp$summary`
## ------------------------------------------------

x <- example_hospital_grp_unit()
x[["summary"]]()
}
\section{Super class}{
\code{\link[HealthDataScotland:health_unitgrp]{HealthDataScotland::health_unitgrp}} -> \code{hospital_grp}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-hospital_grp-plot_types}{\code{hospital_grp$plot_types()}}
\item \href{#method-hospital_grp-plot}{\code{hospital_grp$plot()}}
\item \href{#method-hospital_grp-plot_data}{\code{hospital_grp$plot_data()}}
\item \href{#method-hospital_grp-plot_info}{\code{hospital_grp$plot_info()}}
\item \href{#method-hospital_grp-summary}{\code{hospital_grp$summary()}}
\item \href{#method-hospital_grp-ui}{\code{hospital_grp$ui()}}
\item \href{#method-hospital_grp-server}{\code{hospital_grp$server()}}
\item \href{#method-hospital_grp-clone}{\code{hospital_grp$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="HealthDataScotland" data-topic="health_unitgrp" data-id="ID"><a href='../../HealthDataScotland/html/health_unitgrp.html#method-health_unitgrp-ID'><code>HealthDataScotland::health_unitgrp$ID()</code></a></span></li>
<li><span class="pkg-link" data-pkg="HealthDataScotland" data-topic="health_unitgrp" data-id="IDs"><a href='../../HealthDataScotland/html/health_unitgrp.html#method-health_unitgrp-IDs'><code>HealthDataScotland::health_unitgrp$IDs()</code></a></span></li>
<li><span class="pkg-link" data-pkg="HealthDataScotland" data-topic="health_unitgrp" data-id="addresses"><a href='../../HealthDataScotland/html/health_unitgrp.html#method-health_unitgrp-addresses'><code>HealthDataScotland::health_unitgrp$addresses()</code></a></span></li>
<li><span class="pkg-link" data-pkg="HealthDataScotland" data-topic="health_unitgrp" data-id="combine_data"><a href='../../HealthDataScotland/html/health_unitgrp.html#method-health_unitgrp-combine_data'><code>HealthDataScotland::health_unitgrp$combine_data()</code></a></span></li>
<li><span class="pkg-link" data-pkg="HealthDataScotland" data-topic="health_unitgrp" data-id="data"><a href='../../HealthDataScotland/html/health_unitgrp.html#method-health_unitgrp-data'><code>HealthDataScotland::health_unitgrp$data()</code></a></span></li>
<li><span class="pkg-link" data-pkg="HealthDataScotland" data-topic="health_unitgrp" data-id="datatable"><a href='../../HealthDataScotland/html/health_unitgrp.html#method-health_unitgrp-datatable'><code>HealthDataScotland::health_unitgrp$datatable()</code></a></span></li>
<li><span class="pkg-link" data-pkg="HealthDataScotland" data-topic="health_unitgrp" data-id="download_handler"><a href='../../HealthDataScotland/html/health_unitgrp.html#method-health_unitgrp-download_handler'><code>HealthDataScotland::health_unitgrp$download_handler()</code></a></span></li>
<li><span class="pkg-link" data-pkg="HealthDataScotland" data-topic="health_unitgrp" data-id="health_boards"><a href='../../HealthDataScotland/html/health_unitgrp.html#method-health_unitgrp-health_boards'><code>HealthDataScotland::health_unitgrp$health_boards()</code></a></span></li>
<li><span class="pkg-link" data-pkg="HealthDataScotland" data-topic="health_unitgrp" data-id="health_unit"><a href='../../HealthDataScotland/html/health_unitgrp.html#method-health_unitgrp-health_unit'><code>HealthDataScotland::health_unitgrp$health_unit()</code></a></span></li>
<li><span class="pkg-link" data-pkg="HealthDataScotland" data-topic="health_unitgrp" data-id="initialize"><a href='../../HealthDataScotland/html/health_unitgrp.html#method-health_unitgrp-initialize'><code>HealthDataScotland::health_unitgrp$initialize()</code></a></span></li>
<li><span class="pkg-link" data-pkg="HealthDataScotland" data-topic="health_unitgrp" data-id="metadata"><a href='../../HealthDataScotland/html/health_unitgrp.html#method-health_unitgrp-metadata'><code>HealthDataScotland::health_unitgrp$metadata()</code></a></span></li>
<li><span class="pkg-link" data-pkg="HealthDataScotland" data-topic="health_unitgrp" data-id="subset"><a href='../../HealthDataScotland/html/health_unitgrp.html#method-health_unitgrp-subset'><code>HealthDataScotland::health_unitgrp$subset()</code></a></span></li>
<li><span class="pkg-link" data-pkg="HealthDataScotland" data-topic="health_unitgrp" data-id="titles"><a href='../../HealthDataScotland/html/health_unitgrp.html#method-health_unitgrp-titles'><code>HealthDataScotland::health_unitgrp$titles()</code></a></span></li>
<li><span class="pkg-link" data-pkg="HealthDataScotland" data-topic="health_unitgrp" data-id="validate"><a href='../../HealthDataScotland/html/health_unitgrp.html#method-health_unitgrp-validate'><code>HealthDataScotland::health_unitgrp$validate()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-hospital_grp-plot_types"></a>}}
\if{latex}{\out{\hypertarget{method-hospital_grp-plot_types}{}}}
\subsection{Method \code{plot_types()}}{
Get character vector of plot types for hospital grp.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{hospital_grp$plot_types()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-hospital_grp-plot"></a>}}
\if{latex}{\out{\hypertarget{method-hospital_grp-plot}{}}}
\subsection{Method \code{plot()}}{
Plot hospital grp.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{hospital_grp$plot(type, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{type}}{(character(1))\cr
Character specifying plot type. See \code{plot_types} for options.}

\item{\code{...}}{Passed to plot functions.}
}
\if{html}{\out{</div>}}
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{x <- example_hospital_grp_unit()
x[["plot"]](type = "hospital_bar")
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-hospital_grp-plot_data"></a>}}
\if{latex}{\out{\hypertarget{method-hospital_grp-plot_data}{}}}
\subsection{Method \code{plot_data()}}{
Generate plot data for hospital grp.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{hospital_grp$plot_data(type, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{type}}{(character(1))\cr
Character specifying plot type. See \code{plot_types} for options.}

\item{\code{...}}{Passed to plot data functions.}
}
\if{html}{\out{</div>}}
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{x <- example_hospital_grp_unit()
x[["plot_data"]](type = "hospital_bar")
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-hospital_grp-plot_info"></a>}}
\if{latex}{\out{\hypertarget{method-hospital_grp-plot_info}{}}}
\subsection{Method \code{plot_info()}}{
Get plot info for hospital grp.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{hospital_grp$plot_info(type, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{type}}{(character(1))\cr
Character specifying plot type. See \code{plot_types}
for options.}

\item{\code{...}}{Passed to plot info functions.}
}
\if{html}{\out{</div>}}
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{x <- example_hospital_grp_unit()
x[["plot_info"]](type = "hospital_bar")
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-hospital_grp-summary"></a>}}
\if{latex}{\out{\hypertarget{method-hospital_grp-summary}{}}}
\subsection{Method \code{summary()}}{
Summarise hospital grp data.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{hospital_grp$summary(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Passed to method.}
}
\if{html}{\out{</div>}}
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{x <- example_hospital_grp_unit()
x[["summary"]]()
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-hospital_grp-ui"></a>}}
\if{latex}{\out{\hypertarget{method-hospital_grp-ui}{}}}
\subsection{Method \code{ui()}}{
Create UI for hospital group object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{hospital_grp$ui(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Passed to functions.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-hospital_grp-server"></a>}}
\if{latex}{\out{\hypertarget{method-hospital_grp-server}{}}}
\subsection{Method \code{server()}}{
Create server for hospital group object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{hospital_grp$server()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-hospital_grp-clone"></a>}}
\if{latex}{\out{\hypertarget{method-hospital_grp-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{hospital_grp$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
