% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_download.R
\name{load_processed_data}
\alias{load_processed_data}
\title{Load processed data based on config.}
\usage{
load_processed_data(config = get_config(), ...)
}
\arguments{
\item{config}{Config object specifying whether to use development or
or production protocol for loading data.}

\item{...}{Passed to methods.}
}
\description{
Load processed data based on config.
}
