% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map.R
\name{map}
\alias{map}
\title{R6 class for creating interactive map of health units.}
\description{
R6 class for creating interactive map of health units.

R6 class for creating interactive map of health units.
}
\details{
This R6 class is designed to store sf spatial data for GP practices,
hospitals and health boards. This class is used to create a map
with interactive pins of health centres.
}
\examples{
x <- example_map_object()
x[["id"]]()
x[["title"]]()
x[["sf"]]()
x[["plot"]](type = "interactive_map")
x[["plot_data"]](type = "interactive_map")
x[["plot_info"]](type = "interactive_map")

## ------------------------------------------------
## Method `map$plot`
## ------------------------------------------------

x <- example_map_object()
x[["plot"]](type = "interactive_map")

## ------------------------------------------------
## Method `map$plot_data`
## ------------------------------------------------

x <- example_map_object()
x[["plot_data"]](type = "interactive_map")

## ------------------------------------------------
## Method `map$plot_info`
## ------------------------------------------------

x <- example_map_object()
x[["plot_info"]](type = "interactive_map")
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{.sf}}{A sf spatial data.frame storing spatial data for health units.}

\item{\code{.id}}{Character ID of object.}

\item{\code{.title}}{Character title of object.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-map-new}{\code{map$new()}}
\item \href{#method-map-validate}{\code{map$validate()}}
\item \href{#method-map-sf}{\code{map$sf()}}
\item \href{#method-map-id}{\code{map$id()}}
\item \href{#method-map-title}{\code{map$title()}}
\item \href{#method-map-available_plots}{\code{map$available_plots()}}
\item \href{#method-map-plot}{\code{map$plot()}}
\item \href{#method-map-plot_data}{\code{map$plot_data()}}
\item \href{#method-map-plot_info}{\code{map$plot_info()}}
\item \href{#method-map-ui}{\code{map$ui()}}
\item \href{#method-map-server}{\code{map$server()}}
\item \href{#method-map-clone}{\code{map$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-map-new"></a>}}
\if{latex}{\out{\hypertarget{method-map-new}{}}}
\subsection{Method \code{new()}}{
Create instance of map.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{map$new(.sf, .id = "map", .title = "Interactive Map")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{.sf}}{(\code{sf})\cr
A sf spatial data.frame storing spatial data for health units.}

\item{\code{.id}}{(character(1))\cr
Character id of object. Default is "map".}

\item{\code{.title}}{(character(1))\cr
Character title of object. Default is "Interactive Map".}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-map-validate"></a>}}
\if{latex}{\out{\hypertarget{method-map-validate}{}}}
\subsection{Method \code{validate()}}{
Validate structure of map object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{map$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-map-sf"></a>}}
\if{latex}{\out{\hypertarget{method-map-sf}{}}}
\subsection{Method \code{sf()}}{
Get sf spatial data.frame of map object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{map$sf()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-map-id"></a>}}
\if{latex}{\out{\hypertarget{method-map-id}{}}}
\subsection{Method \code{id()}}{
Get id of map object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{map$id()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-map-title"></a>}}
\if{latex}{\out{\hypertarget{method-map-title}{}}}
\subsection{Method \code{title()}}{
Get title of map object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{map$title()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-map-available_plots"></a>}}
\if{latex}{\out{\hypertarget{method-map-available_plots}{}}}
\subsection{Method \code{available_plots()}}{
Get character vector of available plots for map object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{map$available_plots()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-map-plot"></a>}}
\if{latex}{\out{\hypertarget{method-map-plot}{}}}
\subsection{Method \code{plot()}}{
Plot map object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{map$plot(type, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{type}}{(character(1))\cr
Character specifying plot type. See \code{available_plots}
for options.}

\item{\code{...}}{Passed to plot functions.}
}
\if{html}{\out{</div>}}
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{x <- example_map_object()
x[["plot"]](type = "interactive_map")
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-map-plot_data"></a>}}
\if{latex}{\out{\hypertarget{method-map-plot_data}{}}}
\subsection{Method \code{plot_data()}}{
Generate plot data for map object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{map$plot_data(type, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{type}}{(character(1))\cr
Character specifying plot type. See \code{available_plots}
for options.}

\item{\code{...}}{Passed to plot data functions.}
}
\if{html}{\out{</div>}}
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{x <- example_map_object()
x[["plot_data"]](type = "interactive_map")
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-map-plot_info"></a>}}
\if{latex}{\out{\hypertarget{method-map-plot_info}{}}}
\subsection{Method \code{plot_info()}}{
Get plot infot for map object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{map$plot_info(type, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{type}}{(character(1))\cr
Character specifying plot type. See \code{available_plots}
for options.}

\item{\code{...}}{Passed to plot data functions.}
}
\if{html}{\out{</div>}}
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{x <- example_map_object()
x[["plot_info"]](type = "interactive_map")
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-map-ui"></a>}}
\if{latex}{\out{\hypertarget{method-map-ui}{}}}
\subsection{Method \code{ui()}}{
Create ui for map object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{map$ui()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-map-server"></a>}}
\if{latex}{\out{\hypertarget{method-map-server}{}}}
\subsection{Method \code{server()}}{
Create server for map object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{map$server(data)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{(list)\cr
List of with gp_grp and hospital_grp objects to use for interactive
map pop-up modals.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-map-clone"></a>}}
\if{latex}{\out{\hypertarget{method-map-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{map$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
