% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_download.R
\name{save_processed_data}
\alias{save_processed_data}
\title{Save processed data based on config.}
\usage{
save_processed_data(config = get_config(), ...)
}
\arguments{
\item{config}{Config object specifying whether to use development or
or production protocol for loading data.}

\item{...}{Passed to methods.}
}
\description{
Save processed data based on config.
}
