% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_data.R
\docType{data}
\name{example_gp_data}
\alias{example_gp_data}
\title{Example GP dataset.}
\description{
Data.frame object containing demography data for scottish GP practices
from October 2023 to April 2024.
}
\seealso{
Other health data scotland examples: 
\code{\link{example_gp_metadata}},
\code{\link{example_hospital_data}},
\code{\link{example_hospital_metadata}}
}
\concept{health data scotland examples}
\keyword{data}
