% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_data.R
\docType{data}
\name{example_gp_metadata}
\alias{example_gp_metadata}
\title{Example GP metadata}
\description{
Data.frame object containing  metadata for Scottish GP Practices
from April 2024.
}
\seealso{
Other gp scotland examples: 
\code{\link{example_gp_data}}
}
\concept{gp scotland examples}
\keyword{data}
