% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_processing.R
\name{get_sf}
\alias{get_sf}
\title{Get SF spatial data.frame for selected data type.}
\usage{
get_sf(type = available_sf_types())
}
\arguments{
\item{type}{Character type of spatial data. See \code{available_sf_types}.}
}
\description{
Get SF spatial data.frame for selected data type.
}
