% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gp.R
\name{gp}
\alias{gp}
\title{R6 class storing health statistics for a GP practice.}
\description{
R6 class storing health statistics for a GP practice.

R6 class storing health statistics for a GP practice.
}
\section{Super class}{
\code{\link[HealthDataScotland:health_unit]{HealthDataScotland::health_unit}} -> \code{gp}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-gp-telephone}{\code{gp$telephone()}}
\item \href{#method-gp-gp_cluster}{\code{gp$gp_cluster()}}
\item \href{#method-gp-available_plots}{\code{gp$available_plots()}}
\item \href{#method-gp-plot}{\code{gp$plot()}}
\item \href{#method-gp-plot_data}{\code{gp$plot_data()}}
\item \href{#method-gp-ui}{\code{gp$ui()}}
\item \href{#method-gp-server}{\code{gp$server()}}
\item \href{#method-gp-clone}{\code{gp$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="HealthDataScotland" data-topic="health_unit" data-id="ID"><a href='../../HealthDataScotland/html/health_unit.html#method-health_unit-ID'><code>HealthDataScotland::health_unit$ID()</code></a></span></li>
<li><span class="pkg-link" data-pkg="HealthDataScotland" data-topic="health_unit" data-id="address"><a href='../../HealthDataScotland/html/health_unit.html#method-health_unit-address'><code>HealthDataScotland::health_unit$address()</code></a></span></li>
<li><span class="pkg-link" data-pkg="HealthDataScotland" data-topic="health_unit" data-id="data"><a href='../../HealthDataScotland/html/health_unit.html#method-health_unit-data'><code>HealthDataScotland::health_unit$data()</code></a></span></li>
<li><span class="pkg-link" data-pkg="HealthDataScotland" data-topic="health_unit" data-id="health_board"><a href='../../HealthDataScotland/html/health_unit.html#method-health_unit-health_board'><code>HealthDataScotland::health_unit$health_board()</code></a></span></li>
<li><span class="pkg-link" data-pkg="HealthDataScotland" data-topic="health_unit" data-id="initialize"><a href='../../HealthDataScotland/html/health_unit.html#method-health_unit-initialize'><code>HealthDataScotland::health_unit$initialize()</code></a></span></li>
<li><span class="pkg-link" data-pkg="HealthDataScotland" data-topic="health_unit" data-id="popup_modal"><a href='../../HealthDataScotland/html/health_unit.html#method-health_unit-popup_modal'><code>HealthDataScotland::health_unit$popup_modal()</code></a></span></li>
<li><span class="pkg-link" data-pkg="HealthDataScotland" data-topic="health_unit" data-id="title"><a href='../../HealthDataScotland/html/health_unit.html#method-health_unit-title'><code>HealthDataScotland::health_unit$title()</code></a></span></li>
<li><span class="pkg-link" data-pkg="HealthDataScotland" data-topic="health_unit" data-id="validate"><a href='../../HealthDataScotland/html/health_unit.html#method-health_unit-validate'><code>HealthDataScotland::health_unit$validate()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-gp-telephone"></a>}}
\if{latex}{\out{\hypertarget{method-gp-telephone}{}}}
\subsection{Method \code{telephone()}}{
Get telephone number of GP practice.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{gp$telephone()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-gp-gp_cluster"></a>}}
\if{latex}{\out{\hypertarget{method-gp-gp_cluster}{}}}
\subsection{Method \code{gp_cluster()}}{
Get GP cluster
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{gp$gp_cluster()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-gp-available_plots"></a>}}
\if{latex}{\out{\hypertarget{method-gp-available_plots}{}}}
\subsection{Method \code{available_plots()}}{
Get character vector of available plots for gp unit. Options
are either "population_pyramid" plot or "population_trend" plot.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{gp$available_plots()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-gp-plot"></a>}}
\if{latex}{\out{\hypertarget{method-gp-plot}{}}}
\subsection{Method \code{plot()}}{
Plot gp unit.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{gp$plot(type, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{type}}{(character(1))\cr
Character specifying plot type. See \code{available_plots}
for options.}

\item{\code{...}}{Passed to plot functions.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-gp-plot_data"></a>}}
\if{latex}{\out{\hypertarget{method-gp-plot_data}{}}}
\subsection{Method \code{plot_data()}}{
Generate plot data for gp unit.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{gp$plot_data(type, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{type}}{(character(1))\cr
Character specifying plot type. See \code{available_plots}
for options.}

\item{\code{...}}{Passed to plot_data functions.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-gp-ui"></a>}}
\if{latex}{\out{\hypertarget{method-gp-ui}{}}}
\subsection{Method \code{ui()}}{
Create UI for general practice object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{gp$ui(ns)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{ns}}{Namespace of shiny application page.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-gp-server"></a>}}
\if{latex}{\out{\hypertarget{method-gp-server}{}}}
\subsection{Method \code{server()}}{
Create server for general practice object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{gp$server()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-gp-clone"></a>}}
\if{latex}{\out{\hypertarget{method-gp-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{gp$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
