% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gp_grp.R
\name{gp_grp}
\alias{gp_grp}
\title{R6 class storing health statistics for a list of gp health units.}
\description{
R6 class storing health statistics for a list of gp health units.

R6 class storing health statistics for a list of gp health units.
}
\details{
This R6 class is designed to store population demography data for multiple
GP practices. This class can be used to plot summary statistics and
create shiny UI/server objects.
}
\examples{
gps <- lapply(c("10002", "10017"), example_gp_unit)
x <- gp_grp[["new"]](gps, .id = "gp")
x[["ID"]]()
x[["IDs"]]()
x[["titles"]]()
x[["metadata"]]()
x[["data"]]()
x[["subset"]](id = "10002")
x[["plot"]](type = "national_pyramid")
x[["plot_data"]](type = "national_pyramid")
x[["plot_info"]](type = "national_pyramid")
\dontrun{
x[["ui"]]()
x[["server"]]()
}

## ------------------------------------------------
## Method `gp_grp$plot`
## ------------------------------------------------

x <- example_gp_grp_unit()
x[["plot"]](type = "gp_bar")

## ------------------------------------------------
## Method `gp_grp$plot_data`
## ------------------------------------------------

x <- example_gp_grp_unit()
x[["plot_data"]](type = "gp_bar")

## ------------------------------------------------
## Method `gp_grp$plot_info`
## ------------------------------------------------

x <- example_gp_grp_unit()
x[["plot_info"]](type = "gp_bar")
}
\section{Super class}{
\code{\link[HealthDataScotland:health_unitgrp]{HealthDataScotland::health_unitgrp}} -> \code{gp_grp}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-gp_grp-available_plots}{\code{gp_grp$available_plots()}}
\item \href{#method-gp_grp-plot}{\code{gp_grp$plot()}}
\item \href{#method-gp_grp-plot_data}{\code{gp_grp$plot_data()}}
\item \href{#method-gp_grp-plot_info}{\code{gp_grp$plot_info()}}
\item \href{#method-gp_grp-ui}{\code{gp_grp$ui()}}
\item \href{#method-gp_grp-server}{\code{gp_grp$server()}}
\item \href{#method-gp_grp-clone}{\code{gp_grp$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="HealthDataScotland" data-topic="health_unitgrp" data-id="ID"><a href='../../HealthDataScotland/html/health_unitgrp.html#method-health_unitgrp-ID'><code>HealthDataScotland::health_unitgrp$ID()</code></a></span></li>
<li><span class="pkg-link" data-pkg="HealthDataScotland" data-topic="health_unitgrp" data-id="IDs"><a href='../../HealthDataScotland/html/health_unitgrp.html#method-health_unitgrp-IDs'><code>HealthDataScotland::health_unitgrp$IDs()</code></a></span></li>
<li><span class="pkg-link" data-pkg="HealthDataScotland" data-topic="health_unitgrp" data-id="combine_data"><a href='../../HealthDataScotland/html/health_unitgrp.html#method-health_unitgrp-combine_data'><code>HealthDataScotland::health_unitgrp$combine_data()</code></a></span></li>
<li><span class="pkg-link" data-pkg="HealthDataScotland" data-topic="health_unitgrp" data-id="data"><a href='../../HealthDataScotland/html/health_unitgrp.html#method-health_unitgrp-data'><code>HealthDataScotland::health_unitgrp$data()</code></a></span></li>
<li><span class="pkg-link" data-pkg="HealthDataScotland" data-topic="health_unitgrp" data-id="download_handler"><a href='../../HealthDataScotland/html/health_unitgrp.html#method-health_unitgrp-download_handler'><code>HealthDataScotland::health_unitgrp$download_handler()</code></a></span></li>
<li><span class="pkg-link" data-pkg="HealthDataScotland" data-topic="health_unitgrp" data-id="health_unit"><a href='../../HealthDataScotland/html/health_unitgrp.html#method-health_unitgrp-health_unit'><code>HealthDataScotland::health_unitgrp$health_unit()</code></a></span></li>
<li><span class="pkg-link" data-pkg="HealthDataScotland" data-topic="health_unitgrp" data-id="initialize"><a href='../../HealthDataScotland/html/health_unitgrp.html#method-health_unitgrp-initialize'><code>HealthDataScotland::health_unitgrp$initialize()</code></a></span></li>
<li><span class="pkg-link" data-pkg="HealthDataScotland" data-topic="health_unitgrp" data-id="metadata"><a href='../../HealthDataScotland/html/health_unitgrp.html#method-health_unitgrp-metadata'><code>HealthDataScotland::health_unitgrp$metadata()</code></a></span></li>
<li><span class="pkg-link" data-pkg="HealthDataScotland" data-topic="health_unitgrp" data-id="subset"><a href='../../HealthDataScotland/html/health_unitgrp.html#method-health_unitgrp-subset'><code>HealthDataScotland::health_unitgrp$subset()</code></a></span></li>
<li><span class="pkg-link" data-pkg="HealthDataScotland" data-topic="health_unitgrp" data-id="titles"><a href='../../HealthDataScotland/html/health_unitgrp.html#method-health_unitgrp-titles'><code>HealthDataScotland::health_unitgrp$titles()</code></a></span></li>
<li><span class="pkg-link" data-pkg="HealthDataScotland" data-topic="health_unitgrp" data-id="validate"><a href='../../HealthDataScotland/html/health_unitgrp.html#method-health_unitgrp-validate'><code>HealthDataScotland::health_unitgrp$validate()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-gp_grp-available_plots"></a>}}
\if{latex}{\out{\hypertarget{method-gp_grp-available_plots}{}}}
\subsection{Method \code{available_plots()}}{
Get character vector of available plots for gp grp.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{gp_grp$available_plots()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-gp_grp-plot"></a>}}
\if{latex}{\out{\hypertarget{method-gp_grp-plot}{}}}
\subsection{Method \code{plot()}}{
Plot gp grp.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{gp_grp$plot(type, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{type}}{(character(1))\cr
Character specifying plot type. See \code{available_plots}
for options.}

\item{\code{...}}{Passed to plot functions.}
}
\if{html}{\out{</div>}}
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{x <- example_gp_grp_unit()
x[["plot"]](type = "gp_bar")
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-gp_grp-plot_data"></a>}}
\if{latex}{\out{\hypertarget{method-gp_grp-plot_data}{}}}
\subsection{Method \code{plot_data()}}{
Get plot data for gp grp.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{gp_grp$plot_data(type, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{type}}{(character(1))\cr
Character specifying plot type. See \code{available_plots}
for options.}

\item{\code{...}}{Passed to plot data functions.}
}
\if{html}{\out{</div>}}
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{x <- example_gp_grp_unit()
x[["plot_data"]](type = "gp_bar")
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-gp_grp-plot_info"></a>}}
\if{latex}{\out{\hypertarget{method-gp_grp-plot_info}{}}}
\subsection{Method \code{plot_info()}}{
Get plot info for gp grp.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{gp_grp$plot_info(type, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{type}}{(character(1))\cr
Character specifying plot type. See \code{available_plots}
for options.}

\item{\code{...}}{Passed to plot info functions.}
}
\if{html}{\out{</div>}}
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{x <- example_gp_grp_unit()
x[["plot_info"]](type = "gp_bar")
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-gp_grp-ui"></a>}}
\if{latex}{\out{\hypertarget{method-gp_grp-ui}{}}}
\subsection{Method \code{ui()}}{
Create UI for general practice group object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{gp_grp$ui(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Passed to functions.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-gp_grp-server"></a>}}
\if{latex}{\out{\hypertarget{method-gp_grp-server}{}}}
\subsection{Method \code{server()}}{
Create server for general practice group object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{gp_grp$server()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-gp_grp-clone"></a>}}
\if{latex}{\out{\hypertarget{method-gp_grp-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{gp_grp$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
