% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/health_unitgp.R
\name{health_unitgrp}
\alias{health_unitgrp}
\title{R6 class storing health statistics for a list of health units.}
\description{
R6 class storing health statistics for a list of health units.

R6 class storing health statistics for a list of health units.
}
\examples{

## ------------------------------------------------
## Method `health_unitgrp$summary`
## ------------------------------------------------

x <- example_gp_grp_unit()
x[["summary"]]()
x <- example_hospital_grp_unit()
x[["summary"]]()

## ------------------------------------------------
## Method `health_unitgrp$summary_info`
## ------------------------------------------------

x <- example_gp_grp_unit()
x[["summary_info"]]()
x <- example_hospital_grp_unit()
x[["summary_info"]]()

## ------------------------------------------------
## Method `health_unitgrp$datatable`
## ------------------------------------------------

x <- example_gp_grp_unit()
x[["datatable"]]()
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{.data}}{A list of health units in health unit grp.}

\item{\code{.id}}{A character ID of object.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-health_unitgrp-new}{\code{health_unitgrp$new()}}
\item \href{#method-health_unitgrp-validate}{\code{health_unitgrp$validate()}}
\item \href{#method-health_unitgrp-metadata}{\code{health_unitgrp$metadata()}}
\item \href{#method-health_unitgrp-data}{\code{health_unitgrp$data()}}
\item \href{#method-health_unitgrp-combine_data}{\code{health_unitgrp$combine_data()}}
\item \href{#method-health_unitgrp-ID}{\code{health_unitgrp$ID()}}
\item \href{#method-health_unitgrp-IDs}{\code{health_unitgrp$IDs()}}
\item \href{#method-health_unitgrp-titles}{\code{health_unitgrp$titles()}}
\item \href{#method-health_unitgrp-addresses}{\code{health_unitgrp$addresses()}}
\item \href{#method-health_unitgrp-health_boards}{\code{health_unitgrp$health_boards()}}
\item \href{#method-health_unitgrp-health_unit}{\code{health_unitgrp$health_unit()}}
\item \href{#method-health_unitgrp-subset}{\code{health_unitgrp$subset()}}
\item \href{#method-health_unitgrp-download_handler}{\code{health_unitgrp$download_handler()}}
\item \href{#method-health_unitgrp-summary_types}{\code{health_unitgrp$summary_types()}}
\item \href{#method-health_unitgrp-summary}{\code{health_unitgrp$summary()}}
\item \href{#method-health_unitgrp-summary_info}{\code{health_unitgrp$summary_info()}}
\item \href{#method-health_unitgrp-datatable}{\code{health_unitgrp$datatable()}}
\item \href{#method-health_unitgrp-clone}{\code{health_unitgrp$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-health_unitgrp-new"></a>}}
\if{latex}{\out{\hypertarget{method-health_unitgrp-new}{}}}
\subsection{Method \code{new()}}{
Create instance of health unit grp.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{health_unitgrp$new(.data, .id)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{.data}}{(\code{list})\cr
A list of health units in health unit grp.}

\item{\code{.id}}{(\code{characer})
A character ID of object.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-health_unitgrp-validate"></a>}}
\if{latex}{\out{\hypertarget{method-health_unitgrp-validate}{}}}
\subsection{Method \code{validate()}}{
Validate structure of health unit grp.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{health_unitgrp$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-health_unitgrp-metadata"></a>}}
\if{latex}{\out{\hypertarget{method-health_unitgrp-metadata}{}}}
\subsection{Method \code{metadata()}}{
Get combine metadata of health unit grp.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{health_unitgrp$metadata()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-health_unitgrp-data"></a>}}
\if{latex}{\out{\hypertarget{method-health_unitgrp-data}{}}}
\subsection{Method \code{data()}}{
Get data of health unit grp.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{health_unitgrp$data()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-health_unitgrp-combine_data"></a>}}
\if{latex}{\out{\hypertarget{method-health_unitgrp-combine_data}{}}}
\subsection{Method \code{combine_data()}}{
Get combined metadata and data in single data.frame.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{health_unitgrp$combine_data()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-health_unitgrp-ID"></a>}}
\if{latex}{\out{\hypertarget{method-health_unitgrp-ID}{}}}
\subsection{Method \code{ID()}}{
Get ID of health unit grp.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{health_unitgrp$ID()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-health_unitgrp-IDs"></a>}}
\if{latex}{\out{\hypertarget{method-health_unitgrp-IDs}{}}}
\subsection{Method \code{IDs()}}{
Get character vector of IDs for stored health units
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{health_unitgrp$IDs()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-health_unitgrp-titles"></a>}}
\if{latex}{\out{\hypertarget{method-health_unitgrp-titles}{}}}
\subsection{Method \code{titles()}}{
Get character vector of titles for stored health units
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{health_unitgrp$titles()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-health_unitgrp-addresses"></a>}}
\if{latex}{\out{\hypertarget{method-health_unitgrp-addresses}{}}}
\subsection{Method \code{addresses()}}{
Get character vector of addresses for stored health units
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{health_unitgrp$addresses()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-health_unitgrp-health_boards"></a>}}
\if{latex}{\out{\hypertarget{method-health_unitgrp-health_boards}{}}}
\subsection{Method \code{health_boards()}}{
Get character vector of health boards for stored health units
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{health_unitgrp$health_boards()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-health_unitgrp-health_unit"></a>}}
\if{latex}{\out{\hypertarget{method-health_unitgrp-health_unit}{}}}
\subsection{Method \code{health_unit()}}{
Get stored health unit.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{health_unitgrp$health_unit(id)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id}}{(character(1))\cr
Character specifying ID of object to obtain from group.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-health_unitgrp-subset"></a>}}
\if{latex}{\out{\hypertarget{method-health_unitgrp-subset}{}}}
\subsection{Method \code{subset()}}{
Subset health unit group.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{health_unitgrp$subset(id)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id}}{(character())\cr
Character specifying ID (or IDs) of object to obtain from group.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-health_unitgrp-download_handler"></a>}}
\if{latex}{\out{\hypertarget{method-health_unitgrp-download_handler}{}}}
\subsection{Method \code{download_handler()}}{
Get download handler function of health unit group.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{health_unitgrp$download_handler()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-health_unitgrp-summary_types"></a>}}
\if{latex}{\out{\hypertarget{method-health_unitgrp-summary_types}{}}}
\subsection{Method \code{summary_types()}}{
Get character vector of summary types for grp unit
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{health_unitgrp$summary_types()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-health_unitgrp-summary"></a>}}
\if{latex}{\out{\hypertarget{method-health_unitgrp-summary}{}}}
\subsection{Method \code{summary()}}{
Summarise grp unit data.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{health_unitgrp$summary(type, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{type}}{(character(1))\cr
Character specifying summary type. See \code{summary_types} for options.}

\item{\code{...}}{Passed to method.}
}
\if{html}{\out{</div>}}
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{x <- example_gp_grp_unit()
x[["summary"]]()
x <- example_hospital_grp_unit()
x[["summary"]]()
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-health_unitgrp-summary_info"></a>}}
\if{latex}{\out{\hypertarget{method-health_unitgrp-summary_info}{}}}
\subsection{Method \code{summary_info()}}{
Get summary info for grp unit.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{health_unitgrp$summary_info(type, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{type}}{(character(1))\cr
Character specifying summary type. See \code{summary_types} for options.}

\item{\code{...}}{Passed to summary info functions.}
}
\if{html}{\out{</div>}}
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{x <- example_gp_grp_unit()
x[["summary_info"]]()
x <- example_hospital_grp_unit()
x[["summary_info"]]()
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-health_unitgrp-datatable"></a>}}
\if{latex}{\out{\hypertarget{method-health_unitgrp-datatable}{}}}
\subsection{Method \code{datatable()}}{
Create datatable of gp grp data
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{health_unitgrp$datatable(type, ns = NULL, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{type}}{(character(1))\cr
Character specifying summary type. See \code{summary_types} for options.}

\item{\code{ns}}{Shiny Namespace. Default is NULL.}

\item{\code{...}}{Passed to method.}
}
\if{html}{\out{</div>}}
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{x <- example_gp_grp_unit()
x[["datatable"]]()
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-health_unitgrp-clone"></a>}}
\if{latex}{\out{\hypertarget{method-health_unitgrp-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{health_unitgrp$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
