% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/health_unitgp.R
\name{health_unitgrp}
\alias{health_unitgrp}
\title{R6 class storing health statistics for a list of health units.}
\description{
R6 class storing health statistics for a list of health units.

R6 class storing health statistics for a list of health units.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{.data}}{A list of health units in health unit grp.}

\item{\code{.json}}{A SpatialPointsDataFrame storing spatial information.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-health_unitgrp-new}{\code{health_unitgrp$new()}}
\item \href{#method-health_unitgrp-validate}{\code{health_unitgrp$validate()}}
\item \href{#method-health_unitgrp-data}{\code{health_unitgrp$data()}}
\item \href{#method-health_unitgrp-json}{\code{health_unitgrp$json()}}
\item \href{#method-health_unitgrp-ids}{\code{health_unitgrp$ids()}}
\item \href{#method-health_unitgrp-health_unit}{\code{health_unitgrp$health_unit()}}
\item \href{#method-health_unitgrp-subset}{\code{health_unitgrp$subset()}}
\item \href{#method-health_unitgrp-clone}{\code{health_unitgrp$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-health_unitgrp-new"></a>}}
\if{latex}{\out{\hypertarget{method-health_unitgrp-new}{}}}
\subsection{Method \code{new()}}{
Create instance of health unit grp.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{health_unitgrp$new(.data, .json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{.data}}{(\code{list})\cr
A list of health units in health unit grp.}

\item{\code{.json}}{(\code{SpatialPointsDataFrame})
A SpatialPointsDataFrame storing spatial information for
health units.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-health_unitgrp-validate"></a>}}
\if{latex}{\out{\hypertarget{method-health_unitgrp-validate}{}}}
\subsection{Method \code{validate()}}{
Validate structure of health unit grp.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{health_unitgrp$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-health_unitgrp-data"></a>}}
\if{latex}{\out{\hypertarget{method-health_unitgrp-data}{}}}
\subsection{Method \code{data()}}{
Get data of health unit grp.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{health_unitgrp$data()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-health_unitgrp-json"></a>}}
\if{latex}{\out{\hypertarget{method-health_unitgrp-json}{}}}
\subsection{Method \code{json()}}{
Get JSON of health unit grp.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{health_unitgrp$json()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-health_unitgrp-ids"></a>}}
\if{latex}{\out{\hypertarget{method-health_unitgrp-ids}{}}}
\subsection{Method \code{ids()}}{
Get ids of stored health units
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{health_unitgrp$ids()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-health_unitgrp-health_unit"></a>}}
\if{latex}{\out{\hypertarget{method-health_unitgrp-health_unit}{}}}
\subsection{Method \code{health_unit()}}{
Get stored health unit.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{health_unitgrp$health_unit(id)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id}}{(character(1))\cr
Character specifying ID of object to obtain from group.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-health_unitgrp-subset"></a>}}
\if{latex}{\out{\hypertarget{method-health_unitgrp-subset}{}}}
\subsection{Method \code{subset()}}{
Subset health unit group.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{health_unitgrp$subset(id)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id}}{(character())\cr
Character specifying ID (or IDs) of object to obtain from group.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-health_unitgrp-clone"></a>}}
\if{latex}{\out{\hypertarget{method-health_unitgrp-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{health_unitgrp$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
