% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hospital.R
\name{hospital}
\alias{hospital}
\title{R6 class storing health statistics for a Hospital.}
\description{
R6 class storing health statistics for a Hospital.

R6 class storing health statistics for a Hospital.
}
\section{Super class}{
\code{\link[HealthDataScotland:health_unit]{HealthDataScotland::health_unit}} -> \code{hospital}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-hospital-available_plots}{\code{hospital$available_plots()}}
\item \href{#method-hospital-plot}{\code{hospital$plot()}}
\item \href{#method-hospital-plot_data}{\code{hospital$plot_data()}}
\item \href{#method-hospital-ui}{\code{hospital$ui()}}
\item \href{#method-hospital-server}{\code{hospital$server()}}
\item \href{#method-hospital-clone}{\code{hospital$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="HealthDataScotland" data-topic="health_unit" data-id="address"><a href='../../HealthDataScotland/html/health_unit.html#method-health_unit-address'><code>HealthDataScotland::health_unit$address()</code></a></span></li>
<li><span class="pkg-link" data-pkg="HealthDataScotland" data-topic="health_unit" data-id="data"><a href='../../HealthDataScotland/html/health_unit.html#method-health_unit-data'><code>HealthDataScotland::health_unit$data()</code></a></span></li>
<li><span class="pkg-link" data-pkg="HealthDataScotland" data-topic="health_unit" data-id="health_board"><a href='../../HealthDataScotland/html/health_unit.html#method-health_unit-health_board'><code>HealthDataScotland::health_unit$health_board()</code></a></span></li>
<li><span class="pkg-link" data-pkg="HealthDataScotland" data-topic="health_unit" data-id="id"><a href='../../HealthDataScotland/html/health_unit.html#method-health_unit-id'><code>HealthDataScotland::health_unit$id()</code></a></span></li>
<li><span class="pkg-link" data-pkg="HealthDataScotland" data-topic="health_unit" data-id="initialize"><a href='../../HealthDataScotland/html/health_unit.html#method-health_unit-initialize'><code>HealthDataScotland::health_unit$initialize()</code></a></span></li>
<li><span class="pkg-link" data-pkg="HealthDataScotland" data-topic="health_unit" data-id="popup_modal"><a href='../../HealthDataScotland/html/health_unit.html#method-health_unit-popup_modal'><code>HealthDataScotland::health_unit$popup_modal()</code></a></span></li>
<li><span class="pkg-link" data-pkg="HealthDataScotland" data-topic="health_unit" data-id="title"><a href='../../HealthDataScotland/html/health_unit.html#method-health_unit-title'><code>HealthDataScotland::health_unit$title()</code></a></span></li>
<li><span class="pkg-link" data-pkg="HealthDataScotland" data-topic="health_unit" data-id="validate"><a href='../../HealthDataScotland/html/health_unit.html#method-health_unit-validate'><code>HealthDataScotland::health_unit$validate()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-hospital-available_plots"></a>}}
\if{latex}{\out{\hypertarget{method-hospital-available_plots}{}}}
\subsection{Method \code{available_plots()}}{
Get character vector of available plots for hospital unit. Options
are either "specialty_bar" plot.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{hospital$available_plots()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-hospital-plot"></a>}}
\if{latex}{\out{\hypertarget{method-hospital-plot}{}}}
\subsection{Method \code{plot()}}{
Plot hospital unit.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{hospital$plot(type, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{type}}{(character(1))\cr
Character specifying plot type. See \code{available_plots} for options.}

\item{\code{...}}{Passed to plot functions.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-hospital-plot_data"></a>}}
\if{latex}{\out{\hypertarget{method-hospital-plot_data}{}}}
\subsection{Method \code{plot_data()}}{
Generate plot data for hospital unit.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{hospital$plot_data(type, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{type}}{(character(1))\cr
Character specifying plot type. See \code{available_plots} for options.}

\item{\code{...}}{Passed to plot functions.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-hospital-ui"></a>}}
\if{latex}{\out{\hypertarget{method-hospital-ui}{}}}
\subsection{Method \code{ui()}}{
Create UI for hospital object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{hospital$ui(ns)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{ns}}{Namespace of shiny application page.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-hospital-server"></a>}}
\if{latex}{\out{\hypertarget{method-hospital-server}{}}}
\subsection{Method \code{server()}}{
Create server for hospital object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{hospital$server()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-hospital-clone"></a>}}
\if{latex}{\out{\hypertarget{method-hospital-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{hospital$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
