% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hospital_grp.R
\name{hospital_grp}
\alias{hospital_grp}
\title{R6 class storing health statistics for a list of hospital health units.}
\description{
R6 class storing health statistics for a list of hospital health units.

R6 class storing health statistics for a list of hospital health units.
}
\section{Super class}{
\code{\link[HealthDataScotland:health_unitgrp]{HealthDataScotland::health_unitgrp}} -> \code{hospital_grp}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-hospital_grp-available_plots}{\code{hospital_grp$available_plots()}}
\item \href{#method-hospital_grp-plot}{\code{hospital_grp$plot()}}
\item \href{#method-hospital_grp-plot_data}{\code{hospital_grp$plot_data()}}
\item \href{#method-hospital_grp-summary}{\code{hospital_grp$summary()}}
\item \href{#method-hospital_grp-ui}{\code{hospital_grp$ui()}}
\item \href{#method-hospital_grp-server}{\code{hospital_grp$server()}}
\item \href{#method-hospital_grp-clone}{\code{hospital_grp$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="HealthDataScotland" data-topic="health_unitgrp" data-id="data"><a href='../../HealthDataScotland/html/health_unitgrp.html#method-health_unitgrp-data'><code>HealthDataScotland::health_unitgrp$data()</code></a></span></li>
<li><span class="pkg-link" data-pkg="HealthDataScotland" data-topic="health_unitgrp" data-id="download_handler"><a href='../../HealthDataScotland/html/health_unitgrp.html#method-health_unitgrp-download_handler'><code>HealthDataScotland::health_unitgrp$download_handler()</code></a></span></li>
<li><span class="pkg-link" data-pkg="HealthDataScotland" data-topic="health_unitgrp" data-id="download_server"><a href='../../HealthDataScotland/html/health_unitgrp.html#method-health_unitgrp-download_server'><code>HealthDataScotland::health_unitgrp$download_server()</code></a></span></li>
<li><span class="pkg-link" data-pkg="HealthDataScotland" data-topic="health_unitgrp" data-id="download_ui"><a href='../../HealthDataScotland/html/health_unitgrp.html#method-health_unitgrp-download_ui'><code>HealthDataScotland::health_unitgrp$download_ui()</code></a></span></li>
<li><span class="pkg-link" data-pkg="HealthDataScotland" data-topic="health_unitgrp" data-id="get_download"><a href='../../HealthDataScotland/html/health_unitgrp.html#method-health_unitgrp-get_download'><code>HealthDataScotland::health_unitgrp$get_download()</code></a></span></li>
<li><span class="pkg-link" data-pkg="HealthDataScotland" data-topic="health_unitgrp" data-id="health_unit"><a href='../../HealthDataScotland/html/health_unitgrp.html#method-health_unitgrp-health_unit'><code>HealthDataScotland::health_unitgrp$health_unit()</code></a></span></li>
<li><span class="pkg-link" data-pkg="HealthDataScotland" data-topic="health_unitgrp" data-id="id"><a href='../../HealthDataScotland/html/health_unitgrp.html#method-health_unitgrp-id'><code>HealthDataScotland::health_unitgrp$id()</code></a></span></li>
<li><span class="pkg-link" data-pkg="HealthDataScotland" data-topic="health_unitgrp" data-id="ids"><a href='../../HealthDataScotland/html/health_unitgrp.html#method-health_unitgrp-ids'><code>HealthDataScotland::health_unitgrp$ids()</code></a></span></li>
<li><span class="pkg-link" data-pkg="HealthDataScotland" data-topic="health_unitgrp" data-id="initialize"><a href='../../HealthDataScotland/html/health_unitgrp.html#method-health_unitgrp-initialize'><code>HealthDataScotland::health_unitgrp$initialize()</code></a></span></li>
<li><span class="pkg-link" data-pkg="HealthDataScotland" data-topic="health_unitgrp" data-id="sf"><a href='../../HealthDataScotland/html/health_unitgrp.html#method-health_unitgrp-sf'><code>HealthDataScotland::health_unitgrp$sf()</code></a></span></li>
<li><span class="pkg-link" data-pkg="HealthDataScotland" data-topic="health_unitgrp" data-id="subset"><a href='../../HealthDataScotland/html/health_unitgrp.html#method-health_unitgrp-subset'><code>HealthDataScotland::health_unitgrp$subset()</code></a></span></li>
<li><span class="pkg-link" data-pkg="HealthDataScotland" data-topic="health_unitgrp" data-id="titles"><a href='../../HealthDataScotland/html/health_unitgrp.html#method-health_unitgrp-titles'><code>HealthDataScotland::health_unitgrp$titles()</code></a></span></li>
<li><span class="pkg-link" data-pkg="HealthDataScotland" data-topic="health_unitgrp" data-id="validate"><a href='../../HealthDataScotland/html/health_unitgrp.html#method-health_unitgrp-validate'><code>HealthDataScotland::health_unitgrp$validate()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-hospital_grp-available_plots"></a>}}
\if{latex}{\out{\hypertarget{method-hospital_grp-available_plots}{}}}
\subsection{Method \code{available_plots()}}{
Get character vector of available plots for hospital grp. Options
are either "specialty_bar" or "specialty_line" plot.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{hospital_grp$available_plots()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-hospital_grp-plot"></a>}}
\if{latex}{\out{\hypertarget{method-hospital_grp-plot}{}}}
\subsection{Method \code{plot()}}{
Plot hospital grp.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{hospital_grp$plot(type, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{type}}{(character(1))\cr
Character specifying plot type. See \code{available_plots} for options.}

\item{\code{...}}{Passed to plot functions.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-hospital_grp-plot_data"></a>}}
\if{latex}{\out{\hypertarget{method-hospital_grp-plot_data}{}}}
\subsection{Method \code{plot_data()}}{
Generate plot data for hospital grp.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{hospital_grp$plot_data(type, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{type}}{(character(1))\cr
Character specifying plot type. See \code{available_plots} for options.}

\item{\code{...}}{Passed to plot functions.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-hospital_grp-summary"></a>}}
\if{latex}{\out{\hypertarget{method-hospital_grp-summary}{}}}
\subsection{Method \code{summary()}}{
Summarise hospital grp unit.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{hospital_grp$summary(id = "Scotland national", ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id}}{(character(1))\cr
Character specifying the ID to assign to output data.frame.}

\item{\code{...}}{Passed to plot_data.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-hospital_grp-ui"></a>}}
\if{latex}{\out{\hypertarget{method-hospital_grp-ui}{}}}
\subsection{Method \code{ui()}}{
Create UI for hospital group object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{hospital_grp$ui(ns)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{ns}}{Namespace of shiny application page.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-hospital_grp-server"></a>}}
\if{latex}{\out{\hypertarget{method-hospital_grp-server}{}}}
\subsection{Method \code{server()}}{
Create server for hospital group object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{hospital_grp$server()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-hospital_grp-clone"></a>}}
\if{latex}{\out{\hypertarget{method-hospital_grp-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{hospital_grp$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
